+++ gradle.properties
@@ -1,8 +1,8 @@
 # IntelliJ Platform Artifacts Repositories -> https://plugins.jetbrains.com/docs/intellij/intellij-artifacts.html
 
-pluginGroup = com.github.bkimcode.newpageaction
+pluginGroup = com.github.moreaframework.intellijplugin
 pluginName = Morea
-pluginRepositoryUrl = https://github.com/Spring2023MoreaFrameworkPlugin/NewPageAction
+pluginRepositoryUrl = https://github.com/morea-framework/plugin-intellij
 # SemVer format -> https://semver.org
 pluginVersion = 0.0.2
