+++ src/main/java/com/MoreaFramework/settings/AppSettingsState.java
@@ -1,4 +1,4 @@
-package com.newpageaction.settings;
+package com.MoreaFramework.settings;
 
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.components.PersistentStateComponent;
